// frontend/src/lib/api/models/Workspace.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Knowledge Bank API
 * Modular RAG & Agentic Chatbot API
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkspaceStats } from './WorkspaceStats';
import {
    WorkspaceStatsFromJSON,
    WorkspaceStatsFromJSONTyped,
    WorkspaceStatsToJSON,
    WorkspaceStatsToJSONTyped,
} from './WorkspaceStats';

/**
 * 
 * @export
 * @interface Workspace
 */
export interface Workspace {
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    createdAt?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    updatedAt?: string | null;
    /**
     * 
     * @type {WorkspaceStats}
     * @memberof Workspace
     */
    stats?: WorkspaceStats | null;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    llmProvider?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    embeddingProvider?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Workspace
     */
    ragEngine?: string | null;
}

/**
 * Check if a given object implements the Workspace interface.
 */
export function instanceOfWorkspace(value: object): value is Workspace {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function WorkspaceFromJSON(json: any): Workspace {
    return WorkspaceFromJSONTyped(json, false);
}

export function WorkspaceFromJSONTyped(json: any, ignoreDiscriminator: boolean): Workspace {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'updatedAt': json['updated_at'] == null ? undefined : json['updated_at'],
        'stats': json['stats'] == null ? undefined : WorkspaceStatsFromJSON(json['stats']),
        'llmProvider': json['llm_provider'] == null ? undefined : json['llm_provider'],
        'embeddingProvider': json['embedding_provider'] == null ? undefined : json['embedding_provider'],
        'ragEngine': json['rag_engine'] == null ? undefined : json['rag_engine'],
    };
}

export function WorkspaceToJSON(json: any): Workspace {
    return WorkspaceToJSONTyped(json, false);
}

export function WorkspaceToJSONTyped(value?: Workspace | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'created_at': value['createdAt'],
        'updated_at': value['updatedAt'],
        'stats': WorkspaceStatsToJSON(value['stats']),
        'llm_provider': value['llmProvider'],
        'embedding_provider': value['embeddingProvider'],
        'rag_engine': value['ragEngine'],
    };
}