// frontend/src/lib/api/models/WorkspaceUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Knowledge Bank API
 * Modular RAG & Agentic Chatbot API
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface WorkspaceUpdate
 */
export interface WorkspaceUpdate {
    /**
     * 
     * @type {string}
     * @memberof WorkspaceUpdate
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof WorkspaceUpdate
     */
    description?: string | null;
}

/**
 * Check if a given object implements the WorkspaceUpdate interface.
 */
export function instanceOfWorkspaceUpdate(value: object): value is WorkspaceUpdate {
    return true;
}

export function WorkspaceUpdateFromJSON(json: any): WorkspaceUpdate {
    return WorkspaceUpdateFromJSONTyped(json, false);
}

export function WorkspaceUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkspaceUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function WorkspaceUpdateToJSON(json: any): WorkspaceUpdate {
    return WorkspaceUpdateToJSONTyped(json, false);
}

export function WorkspaceUpdateToJSONTyped(value?: WorkspaceUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
    };
}